// maplibre/src/client/models/GeoFeature_PointGeometry_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { PointGeometry } from './PointGeometry';

export type GeoFeature_PointGeometry_ = {
  type?: string;
  geometry: PointGeometry;
  properties: Record<string, any>;
};