// maplibre/src/client/models/PaginatedResponse_ConnectionRequestApiSchema_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ConnectionRequestApiSchema } from './ConnectionRequestApiSchema';

export type PaginatedResponse_ConnectionRequestApiSchema_ = {
  /**
   * Number of items returned in the response
   */
  count: number;
  /**
   * List of items returned in the response following given criteria
   */
  items: Array<ConnectionRequestApiSchema>;
};